ComboBox skin.
I want to change the look/feel of the ComboBox control, therefore I have
to edit the Microsoft source xaml for the default. Ludicrously harsh.

I've done as good a job as I can without investing the time required to
become a designer. I still don't like how it doesn't match the glass
effect on the buttons.

But the plan is to get a designer to reskin the UI theme later anyway.
Still annoying though.