Charting added. Player module finished :)
Not 100% happy with the axis at the bottom of the chart, and I wouldn't
mind the option of a scrollbar like the DataGrids have.

The UI feels potentially a little clunky, a little click-happy. Might
have to revisit it later.

But otherwise, I'm happy to be finished with the most complicated module
thus far. And I thought the Team module was tough.